% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscorev.R
\name{mscorev}
\alias{mscorev}
\title{Computes M-scores for M-tests and estimates.}
\usage{
mscorev(ymat, inner = 0, trim = 2.5, qu = 0.5, TonT = FALSE)
}
\arguments{
\item{ymat}{ymat is a matrix as described in the documentation for senmk().}

\item{inner}{inner is the parameter described in the documentation for senmk().
The inner parameter is discussed in Rosenbaum (2013).}

\item{trim}{trim is the parameter described in the documentation for senmk().
Note that the default for mscorev is trim = 2.5, but other functions in this package reset the default to trim = 3.
This is for consistency with the sensitivitymv package which also contains the mscorev function.}

\item{qu}{qu is the lambda parameter described in the documentation for senmk().}

\item{TonT}{If TonT=FALSE, then the total score in set (row) i is divided by the number ni of individuals in row i, as in expression (8) in Rosenbaum (2007).
If TonT=TRUE, then the division is by ni-1, not by ni, and there is a further division by the total number of matched sets.
See the discussion of TonT in the documentation for senmk().}
}
\value{
Generally, a matrix with the same dimensions as ymat containing the M-scores.
Exception: if a matched set does not contain at least one treated subject and at least one control, then that set will not appear in the result, and the result will have fewer rows than ymat.
However, if a matched set has several controls but no treated subject, then these controls will contribute to the estimate of the scale parameter, typically the median absolute pair difference.
}
\description{
Of limited interest to most users, function mscorev() computes the M-scores used by functions senm(), senmCI(), comparison(), and principal() that perform Huber-Maritz M-tests. The function is also in the package sensitivitymv.
}
