% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separable1vk.R
\name{separable1vk}
\alias{separable1vk}
\title{Asymptotic separable calculations internal to other functions.}
\usage{
separable1vk(ymat, gamma = 1, k)
}
\arguments{
\item{ymat}{ymat is a matrix whose rows are matched sets and whose columns are matched individuals. The first column describes treated individuals. Other columns describe controls. If matched sets contain variable numbers of controls, NAs fill in empty spaces in ymat; see the documentation for senmv. In senmk, the matrix ymat is created by mscorev. Instead, if there were no NAs and ranks within rows were used in ymat, then separable1v would perform a sensitivity analysis for the stratified Wilcoxon two-sample test. Applied directly to data, it performs a sensitivity analysis for the permutational t-test.}

\item{gamma}{gamma is the value of the sensitivity parameter; see the documentation for the senmk function. One should use a value of gamma >= 1.}

\item{k}{k smallest biases are bounded by gamma}
}
\value{
A list.
\itemize{
\item pval - Approximate upper bound on the one-sided P-value.
\item deviate - Deviate that is compared to the upper tail of the standard Normal distribution to obtain the P-value.
\item statistic - Value of the test statistic.
\item expectation - Maximum null expectation of the test statistic for the given value of gamma.
\item variance	- Among null distributions that yield the maximum expectation, variance is the maximum possible variance for the given value of gamma.
}
}
\description{
Of limited interest to most users, this general purpose function is internal to other functions in the sensitivityq package. It is based on separable1v function in the sensitivitymv package, version 1.3. The function performs the asymptotic separable calculations with the loosen precondition that only parts of the biases are bounded.
}
